+++ apps/functions-subscription/src/adapters/azure/cosmosdb/subscription-history.ts
@@ -23,7 +23,7 @@ export const makeSubscriptionHistoryCosmosContainer = (
             container.items
               .query({
                 query:
-                  'SELECT * FROM c WHERE subscriptionId = @sId ORDER BY version DESC OFFEST 0 LIMIT 1',
+                  'SELECT * FROM c WHERE c.subscriptionId = @sId ORDER BY c.version DESC OFFSET 0 LIMIT 1',
                 parameters: [
                   {
                     name: '@sId',