+++ apps/functions-subscription/src/use-cases/__tests__/process-activation-request.test.ts
@@ -10,13 +10,27 @@ import {
   aSubscriptionHistory,
   aSubscriptionHistoryV1,
   anActivationRequest,
+  anActivationRequestActivated,
 } from '../../domain/__tests__/data';
 
 describe('processActivationRequest', () => {
-  it('should create a new version of subscription-history', async () => {
+  it('should return none if activation request is not activated', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
 
+    const actual =
+      await processActivationRequest(anActivationRequest)(testEnv)();
+
+    expect(actual).toStrictEqual(E.right(O.none));
+    expect(mockEnv.subscriptionHistoryReader.getLatest).toBeCalledTimes(0);
+    expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(0);
+  });
+
+  it('should create a new version of subscription-history if activated', async () => {
+    const mockEnv = makeTestEnv();
+    const testEnv = mockEnv as unknown as Capabilities;
+
+    mockEnv.clock.now.mockReturnValueOnce(aSubscriptionHistoryV1.updatedAt);
     mockEnv.hashFn
       .mockReturnValueOnce({ value: aSubscription.id })
       .mockReturnValueOnce({ value: aSubscriptionHistoryV1.id });
@@ -24,13 +38,14 @@ describe('processActivationRequest', () => {
       TE.right(O.some(aSubscriptionHistory)),
     );
     mockEnv.subscriptionHistoryWriter.insert.mockReturnValueOnce(
-      TE.right(void 0),
+      TE.right(aSubscriptionHistoryV1),
     );
 
-    const actual =
-      await processActivationRequest(anActivationRequest)(testEnv)();
+    const actual = await processActivationRequest(anActivationRequestActivated)(
+      testEnv,
+    )();
 
-    expect(actual).toStrictEqual(E.right(void 0));
+    expect(actual).toStrictEqual(E.right(O.some(aSubscriptionHistoryV1)));
     expect(mockEnv.subscriptionHistoryReader.getLatest).toBeCalledWith({
       subscriptionId: aSubscription.id,
     });
@@ -48,8 +63,9 @@ describe('processActivationRequest', () => {
       TE.right(O.none),
     );
 
-    const actual =
-      await processActivationRequest(anActivationRequest)(testEnv)();
+    const actual = await processActivationRequest(anActivationRequestActivated)(
+      testEnv,
+    )();
     const expected = E.left(new Error('Subscription History not found'));
 
     expect(actual).toStrictEqual(expected);
@@ -72,8 +88,9 @@ describe('processActivationRequest', () => {
       TE.left(error),
     );
 
-    const actual =
-      await processActivationRequest(anActivationRequest)(testEnv)();
+    const actual = await processActivationRequest(anActivationRequestActivated)(
+      testEnv,
+    )();
 
     expect(actual).toStrictEqual(E.left(error));
   });
@@ -90,8 +107,9 @@ describe('processActivationRequest', () => {
       TE.left(error),
     );
 
-    const actual =
-      await processActivationRequest(anActivationRequest)(testEnv)();
+    const actual = await processActivationRequest(anActivationRequestActivated)(
+      testEnv,
+    )();
 
     expect(actual).toStrictEqual(E.left(error));
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(0);